backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
anthropic>=0.40.0
python-docx>=0.8.11
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
ifcopenshell>=0.8.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openpyxl>=3.1.0
